% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asf_products}
\alias{asf_products}
\title{Risk of ASF introduction via sub products}
\format{
List with three data frames
A list with the nodes, edges and stratified values to use by the quantrra library to perform risk assessment in R
\describe{
  \item{nodes}{This table contains the nodes form the risk assessment scenario tree and its asociated parameters. The table is used for runing the model with the function quantrra::ra_run()}
  \item{edges}{This table contains the information of the connections between the nodes of the risk assessment tree. The table is used for visualization purposes with the function quantrra::ra_plot_tree()}
  \item{stratified}{This table contains the parameters used to run a stratified analysis with the function ra_run_stratified(). Each row represents one strata and the columns represent the parameters from the model for that strata.}
  ...
}
}
\source{
<https://doi.org/10.1016/j.rvsc.2023.104990>
}
\usage{
asf_products
}
\description{
Risk of ASF introduction via sub products
}
\keyword{datasets}
