% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASFm}
\alias{ASFm}
\title{Risk of ASF introduction via live products}
\format{
## List with two data frames
A list with the nodes and edges to use by the QuantRRA library
\describe{
  \item{nodes}{node table}
  \item{edges}{edges}
  ...
}
}
\source{
<https://dx.plos.org/10.1371/journal.pone.0182850>
}
\usage{
ASFm
}
\description{
Model adapted from the publication by Hernandez-Ibata el. al 2017. 
Report ...
}
\keyword{datasets}
