% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OIRSA}
\alias{OIRSA}
\title{Risk of ASF introduction via live products}
\format{
List with two data frames
A list with the nodes and edges to use by the quantrra library
\describe{
  \item{nodes}{This table contains the nodes form the risk assessment scenario tree and its asociated parameters. The table is used for runing the model with the function quantrra::ra_run()}
  \item{edges}{This table contains the information of the connections between the nodes of the risk assessment tree. The table is used for visualization purposes with the function quantrra::ra_plot_tree()}
  ...
}
#' @source {OIRSA. 2020. Análisis de riesgo sobre la probabilidad de ingreso, establecimiento y diseminación de la PPA en la porcicultura de los países de la región del OIRSA.}\doi{https://www.oirsa.org/contenido/2020/AR_PPA_Edición\%20revisada\%2001_07_20.pdf}
}
\usage{
OIRSA
}
\description{
Model adapted from the publication by OIRSA 2020 to evaluate the risk of introduction of ASF in the OIRSA region (Mexico and Central America).
}
\keyword{datasets}
