% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_of.R
\name{one_of}
\alias{one_of}
\title{Function to Sample from a multinomial distribution}
\usage{
one_of(x, p = NA, n = 1)
}
\arguments{
\item{x}{vector of categories}

\item{p}{vector of probabilities}

\item{n}{Number of samples}
}
\description{
Function to Sample from a multinomial distribution
}
