% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin}
\alias{pin}
\title{Function to estimate the probability of an animal not being detected with a diagnostic test}
\usage{
pin(p0, Se, Sp)
}
\arguments{
\item{p0}{estimated prevalence}

\item{Se}{Test sensitivity}

\item{Sp}{Test specificity}
}
\description{
Function to estimate the probability of an animal not being detected with a diagnostic test
}
