% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.R
\name{plotDist}
\alias{plotDist}
\title{Function to plot a distribution with descriptive statistics}
\usage{
plotDist(x, col = "red4", main = "")
}
\arguments{
\item{x}{Vector of the distribution}
}
\description{
Function to plot a distribution with descriptive statistics
}
