% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_LiveAnimals.R
\name{pt_LiveAnimals}
\alias{pt_LiveAnimals}
\title{Function to estimate the probability of importing a case via live animals}
\usage{
pt_LiveAnimals(pars)
}
\arguments{
\item{pars}{List of parameters for the model}
}
\description{
Function to estimate the probability of importing a case via live animals
}
