% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_MeatProducts.R
\name{pt_MeatProducts}
\alias{pt_MeatProducts}
\title{Function to estimate the probability of importing a case via meat products}
\usage{
pt_MeatProducts(pars)
}
\arguments{
\item{pars}{List of parameters for the model}
}
\description{
Function to estimate the probability of importing a case via meat products
}
