% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_Products.R
\name{pt_Products}
\alias{pt_Products}
\title{Function to estimate the probability of importing a case via other products}
\usage{
pt_Products(pars)
}
\arguments{
\item{pars}{List of parameters for the model}
}
\description{
Function to estimate the probability of importing a case via other products
}
