% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_export.R
\name{ra_export}
\alias{ra_export}
\title{Export a model file to a file for external use}
\usage{
ra_export(m, dir = "model", file = "model", format = "xlsx")
}
\arguments{
\item{m}{Model table, can be a single data frame or a list}

\item{dir}{directory path for the file}

\item{format}{the format used to export the model file, can be: "zip" or "xlsx"}

\item{name}{name of the output file (not necessary to add extension)}
}
\value{
Function will export a file to the specified location on the operating system.
}
\description{
This function conveniently exports a set of nodes and edges into a xlsx or zip file for external use.
The file exported can be used for editing outside R, for the shiny interactive interface of quantrra or imported back into an R session.
}
\examples{
m <- quantrra::ASFm # use one of the included example model files

# This will export only the nodes (model table)
ra_export(m = m$nodes, format = "zip")

# We can also add an edge table, used for visualization purposes
ra_export(m = m, format = "xlsx")

}
