% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_gsa.R
\name{ra_gsa}
\alias{ra_gsa}
\title{Function to perform a  global sensitivity analysis using random forest.}
\usage{
ra_gsa(data, f, main = "", seed = 1, palette = "-RdYlGn", tree = "none")
}
\arguments{
\item{data}{data set for the analysis}

\item{f}{formula}

\item{seed}{seed used for replication purposes}

\item{palette}{Color palette for the static tree}

\item{tree}{Type of tree. options include: 'none' no tree, 'interactive' interactive visualization with visNetwork, and 'static' static plot with rpart.plot}

\item{mainMain}{title for the plots}
}
\description{
This function is based on the publication: Global sensitivity analysis for complex ecological models: a case study of riparian cottonwood population dynamics by Harper et. al 2011.
This method combines Random Forest and CART to rank the most influential parameters in the main outcome and provide a graphic representation of the interaction between the parameters in the outcom.
}
