% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_import.R
\name{ra_import}
\alias{ra_import}
\title{Import a model file from zip or xlsx}
\usage{
ra_import(p)
}
\arguments{
\item{p}{Path to the model file created in the format from ra_export, which can be either a zip file or a xlsx}
}
\description{
This function imports a previously created model file. This can be either a zip or xlsx file.
}
