% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_one_of.R
\name{ra_one_of}
\alias{ra_one_of}
\title{Sample from a multinomial (discrete) distribution}
\usage{
ra_one_of(x, p = NULL, n = 1)
}
\arguments{
\item{x}{vector of categories}

\item{p}{vector of probabilities}

\item{n}{Number of samples}
}
\value{
A vector of values sampled form the distribution specified
}
\description{
This function samples a value from a discrete distribution.
}
\examples{
ra_one_of(x = c("one", "two", "3"), p = c(0.2, 0.3, 0.5), n = 5)

}
