% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_oneOf.R
\name{ra_oneOf}
\alias{ra_oneOf}
\title{Function to Sample from a multinomial distribution}
\usage{
ra_oneOf(x, p = NA, n = 1)
}
\arguments{
\item{x}{vector of categories}

\item{p}{vector of probabilities}

\item{n}{Number of samples}
}
\description{
Function to Sample from a multinomial distribution
}
