% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_plot_dist.R
\name{ra_plot_dist}
\alias{ra_plot_dist}
\title{Plot a distribution with descriptive statistics}
\usage{
ra_plot_dist(x, col = "red4", main = "")
}
\arguments{
\item{x}{Vector of the sampled values}

\item{col}{color used for the fill}

\item{main}{Main text used as a title for the plot}
}
\description{
Plot a distribution with descriptive statistics
}
