% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_plot_ranking.R
\name{ra_plot_ranking}
\alias{ra_plot_ranking}
\title{Function to plot the ranking}
\usage{
ra_plot_ranking(d, var, id = "IDs", col = "red4")
}
\arguments{
\item{var}{variable to plot}

\item{id}{Name of IDs}

\item{col}{color of the plot}

\item{M}{Model outputs}
}
\description{
Function to plot the ranking
}
