% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_plot_score.R
\name{ra_plot_score}
\alias{ra_plot_score}
\title{Function to Fit a distribution}
\usage{
ra_plot_score(x, main = "", brks = NULL)
}
\arguments{
\item{x}{vector of numbers}

\item{main}{Main title for the figure}

\item{brks}{breaks for the risk thresholds}
}
\description{
Function to Fit a distribution
}
