% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_plot_stratified.R
\name{ra_plot_stratified}
\alias{ra_plot_stratified}
\title{Barplot used to rank the estimate and 95 percentile of the output}
\usage{
ra_plot_stratified(
  x,
  var,
  id = "ids",
  col = "red4",
  format = "static",
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{x}{output from a stratified analysis obtained by the function ra_run_strat}

\item{var}{variable to plot}

\item{id}{Name given to the unique identifier of the stratified var. defaul is id = "IDs"}

\item{col}{color of the plot}

\item{format}{whether the figure is static (using ggplot) or interactive (using plotly)}

\item{xlab}{label used for the x axis}

\item{ylab}{label used for the y axis}
}
\value{
if type = static, a ggplot figure with the ranked strata will be returned, otherwise the a plotly figure will be returned
}
\description{
A barplot arranged in descending order is plotted for the estimates and 95% percentiles obtained from a stratified analysis.
}
\examples{
# use one of the examples
m <- quantrra::asf_products
# run the stratified analysis
os <- ra_run_strat(m = m$nodes, nsim = 5e3)
# Plot the ranking
ra_plot_ranking(x = os, var = "Pf")

}
