% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_plotDist.R
\name{ra_plotDist}
\alias{ra_plotDist}
\title{Function to plot a distribution with descriptive statistics}
\usage{
ra_plotDist(x, col = "red4", main = "")
}
\arguments{
\item{x}{Vector of the distribution}
}
\description{
Function to plot a distribution with descriptive statistics
}
