% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_plotScore.R
\name{ra_plotScore}
\alias{ra_plotScore}
\title{Function to Fit a distribution}
\usage{
ra_plotScore(x, main = "", brks = NULL)
}
\arguments{
\item{x}{vector of numbers}

\item{main}{Main title for the figure}

\item{brks}{breaks for the risk thresholds}
}
\description{
Function to Fit a distribution
}
