% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_run.R
\name{ra_run}
\alias{ra_run}
\title{Function to run the stochastic probabilistic model using a QuantRRA table}
\usage{
ra_run(M, nsim)
}
\arguments{
\item{M}{Model file}

\item{nsim}{Number of simulations}
}
\description{
Function to run the stochastic probabilistic model using a QuantRRA table
}
