% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_run_strat.R
\name{ra_run_strat}
\alias{ra_run_strat}
\title{Function to calculate the stratified risk}
\usage{
ra_run_strat(m, tbl, nsim, full = F, simplify = T)
}
\arguments{
\item{m}{Model file}

\item{tbl}{Table with parameters per strata}

\item{nsim}{Number of simulations}
}
\value{
data frame with median, and 95% percentiles for each of the output nodes
}
\description{
This function runs a stratified model
}
