% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_which.R
\name{ra_which}
\alias{ra_which}
\title{Function to sample from different distributions}
\usage{
ra_which(x, distributions)
}
\arguments{
\item{x}{Data to fit}

\item{dist}{Distribution specification}
}
\description{
Function to sample from different distributions
}
