% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_plot.R
\name{ranking_plot}
\alias{ranking_plot}
\title{Function to plot the ranking}
\usage{
ranking_plot(d, var, id = "IDs", col = "red4")
}
\arguments{
\item{var}{variable to plot}

\item{id}{Name of IDs}

\item{col}{color of the plot}

\item{M}{Model outputs}
}
\description{
Function to plot the ranking
}
