% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankingPlot.R
\name{RankingPlot}
\alias{RankingPlot}
\title{Function to plot the ranking}
\usage{
RankingPlot(d, var, id = "IDs", col = "red4")
}
\arguments{
\item{var}{variable to plot}

\item{id}{Name of IDs}

\item{col}{color of the plot}

\item{M}{Model outputs}
}
\description{
Function to plot the ranking
}
