% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCART.R
\name{RFCART}
\alias{RFCART}
\title{Function to do the RF CART}
\usage{
RFCART(data, f, main = "", seed = 1, palette = "-RdYlGn", tree = "none")
}
\arguments{
\item{data}{data set for the analysis}

\item{f}{formula}

\item{seed}{seed used for replication purposes}

\item{palette}{Color palette for the static tree}

\item{tree}{Type of tree. options include: 'none' no tree, 'interactive' interactive visualization with visNetwork, and 'static' static plot with rpart.plot}

\item{mainMain}{title for the plots}
}
\description{
Function to do the RF CART
}
