% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra.R
\name{rra}
\alias{rra}
\title{Function to run the stochastic probabilistic model using a QuantRRA table}
\usage{
rra(M, nsim)
}
\arguments{
\item{M}{Model file}

\item{nsim}{Number of simulations}
}
\description{
Function to run the stochastic probabilistic model using a QuantRRA table
}
