% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichDist.R
\name{whichDist}
\alias{whichDist}
\title{Function to sample from different distributions}
\usage{
whichDist(x, distributions)
}
\arguments{
\item{x}{Data to fit}

\item{dist}{Distribution specification}
}
\description{
Function to sample from different distributions
}
