% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wm}
\alias{wm}
\title{World map}
\format{
simple features MULTIPOLIGON

\describe{
  \item{name}{Name of the country}
  ...
}
}
\usage{
wm
}
\description{
World map in simple feature format. adapted from the rnaturalearth library for visualization purposes
}
\keyword{datasets}
