% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asf_products}
\alias{asf_products}
\title{Risk of ASF introduction via sub products}
\format{
List with two data frames
A list with the model and stratified values to use by the quantrra library to perform risk assessment in R
\describe{
  \item{model}{This table contains the nodes form the risk assessment scenario tree and its asociated parameters. The table is used for runing the model with the function quantrra::ra_run()}
  \item{stratified}{This table contains the parameters used to run a stratified analysis with the function ra_run_stratified(). Each row represents one strata and the columns represent the parameters from the model for that strata.}
  ...
}
}
\source{
{Muñoz-Pérez, Carolina, Beatriz Martínez-López, José Pablo Gómez-Vázquez, Cecilia Aguilar-Vega, Jaime Bosch, Satoshi Ito, Marta Martínez-Avilés, and José Manuel Sánchez-Vizcaíno. 2023. “Quantitative Risk Assessment of African Swine Fever Introduction into Spain by Legal Import of Swine Products.” Research in Veterinary Science, August, 104990.} \doi{https://doi.org/10.1016/j.rvsc.2023.104990}
}
\usage{
asf_products
}
\description{
Model used for the publication: “Quantitative Risk Assessment of African Swine Fever Introduction into Spain by Legal Import of Swine Products.” This model evaluates the risk of introduction of African Swine Fever into Spain via legal product imports.
}
\keyword{datasets}
