% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asf_products}
\alias{asf_products}
\title{Risk of ASF introduction via sub products}
\format{
## List with three data frames
A list with the nodes, edges and stratified values to use by the quantrra library to perform risk assessment in R
\describe{
  \item{nodes}{node table}
  \item{edges}{edges}
  \item{stratified}{}
  ...
}
}
\source{
<https://doi.org/10.1016/j.rvsc.2023.104990>
}
\usage{
asf_products
}
\description{
Risk of ASF introduction via sub products
}
\keyword{datasets}
