% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASFm}
\alias{ASFm}
\title{Risk of ASF introduction via live products}
\format{
List with two data frames
A list with the nodes and edges to use by the quantrra library
\describe{
  \item{nodes}{This table contains the nodes form the risk assessment scenario tree and its asociated parameters. The table is used for runing the model with the function quantrra::ra_run()}
  \item{edges}{This table contains the information of the connections between the nodes of the risk assessment tree. The table is used for visualization purposes with the function quantrra::ra_plot_tree()}
  ...
}
}
\source{
<https://dx.plos.org/10.1371/journal.pone.0182850>
}
\usage{
ASFm
}
\description{
Model adapted from the publication by Hernandez-Ibata el. al 2017.
}
\keyword{datasets}
