% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASFm}
\alias{ASFm}
\title{Risk of ASF introduction via live products}
\format{
List with two data frames
A list with the nodes and edges to use by the quantrra library
\describe{
  \item{model}{This table contains the nodes form the risk assessment scenario tree and its asociated parameters. The table is used for runing the model with the function quantrra::ra_run()}
  ...
}
}
\source{
{Herrera-Ibatá, Diana María, Beatriz Martínez-López, Darla Quijada, Kenneth Burton, and Lina Mur. 2017. “Quantitative Approach for the Risk Assessment of African Swine Fever and Classical Swine Fever Introduction into the United States Through Legal Imports of Pigs and Swine Products.” Edited by Shuo Su. PLOS ONE 12 (8): e0182850.} \doi{https://doi.org/10.1371/journal.pone.0182850}
}
\usage{
ASFm
}
\description{
Model adapted from the publication by Hernandez-Ibata el. al 2017.
}
\keyword{datasets}
