% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{FitDist}
\alias{FitDist}
\title{Function to Fit a distribution}
\usage{
FitDist(x, dist, full = F, bw = 0.2, digits = 2)
}
\arguments{
\item{x}{Data to fit}

\item{dist}{Distribution specification}
}
\description{
Function to Fit a distribution
}
