% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTree.R
\name{ModelTree}
\alias{ModelTree}
\title{Function to plot a model tree from a QuantRRA table}
\usage{
ModelTree(M)
}
\arguments{
\item{M}{Model list with nodes and edges, see example QuantRRA::ASFm}
}
\description{
Function to plot a model tree from a QuantRRA table
}
