% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OIRSA}
\alias{OIRSA}
\title{Risk of ASF introduction via live products}
\format{
data.frame in the format of the model table
A data.frame with the variables needed by the quantrra library to run a risk assessment model
\describe{
  \item{model}{This table contains the nodes form the risk assessment scenario tree and its asociated parameters. The table is used for runing the model with the function quantrra::ra_run()}
  ...
}
}
\source{
{OIRSA. 2020. Análisis de riesgo sobre la probabilidad de ingreso, establecimiento y diseminación de la PPA en la porcicultura de los países de la región del OIRSA.}
}
\usage{
OIRSA
}
\description{
Model adapted from the publication by OIRSA 2020 to evaluate the risk of introduction of ASF in the OIRSA region (Mexico and Central America).
}
\keyword{datasets}
