% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OIRSA}
\alias{OIRSA}
\title{Risk of ASF introduction via live products}
\format{

}
\usage{
OIRSA
}
\description{
Model adapted from the publication by OIRSA 2020 to evaluate the risk of introduction of ASF in the OIRSA region (Mexico and Central America).
}
\keyword{datasets}
