% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_export.R
\name{ra_export}
\alias{ra_export}
\title{Function to export a model file}
\usage{
ra_export(n, e, dir = "model", name = "model")
}
\arguments{
\item{n}{table of nodes}

\item{e}{table of edges}

\item{dir}{directory path for the file}

\item{name}{name of the output file (not necessary to ad extension .zip)}
}
\description{
Function to export a model file
}
