% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_export.R
\name{ra_export}
\alias{ra_export}
\title{Export a model file to a zip file for external use}
\usage{
ra_export(n, e, dir = "model", name = "model")
}
\arguments{
\item{n}{table of nodes}

\item{e}{table of edges}

\item{dir}{directory path for the file}

\item{name}{name of the output file (not necessary to ad extension .zip)}
}
\description{
This function conveniently exports a set of nodes and edges into a zip file for external use.
The zip file exported can be used in the shiny interactive interface of quantrra or imported back into an R session.
}
