% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_fit.R
\name{ra_fit}
\alias{ra_fit}
\title{(experimental) fit a distribution}
\usage{
ra_fit(x, dist, full = F, bw = 0.2, digits = 2)
}
\arguments{
\item{x}{Data to fit}

\item{dist}{Distribution specification}
}
\description{
Experimental: This function fits a vector of values to a specified distribution. 
The distribution is fitted based on an optimization of the residual sum of squares (rss).
Currently, the function supports only a few distributions, including: NORMAL, POISSON, PERT, TRIANGLE, and UNIFORM
}
