% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_gsa.R
\name{ra_gsa}
\alias{ra_gsa}
\title{Function to perform a  global sensitivity analysis using random forest.}
\usage{
ra_gsa(data, f, main = "", seed = 1, palette = "-RdYlGn", tree = "none")
}
\arguments{
\item{data}{data set for the analysis}

\item{f}{formula}

\item{seed}{seed used for replication purposes}

\item{palette}{Color palette for the static tree}

\item{tree}{Type of tree. options include: 'none' no tree, 'interactive' interactive visualization with visNetwork, and 'static' static plot with rpart.plot}

\item{mainMain}{title for the plots}
}
\value{
A list with the following elements: $VarianceExp The variance explained by the GSA, $RelImport a plot ranking the variables by their relative importance
}
\description{
This function is based on the publication: Global sensitivity analysis for complex ecological models: a case study of riparian cottonwood population dynamics by Harper et. al 2011.
This method combines Random Forest and CART to rank the most influential parameters in the main outcome and provide a graphic representation of the interaction between the parameters in the outcom.
}
\examples{
set.seed(1)

# use one of the example models
m <- quantrra::OIRSA

# run the model
output <- ra_run(m = m$nodes, nsim = 1000)

# Run the GSA on the model output
ra_gsa(
  data = output,
  f = P ~ P1 + P2 + P3 + R1 + R2
)

}
