% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_pin.R
\name{ra_pin}
\alias{ra_pin}
\title{Estimate the probability of an animal not being detected with a diagnostic test}
\usage{
ra_pin(p0, Se, Sp)
}
\arguments{
\item{p0}{estimated prevalence}

\item{Se}{Test sensitivity}

\item{Sp}{Test specificity}
}
\description{
Estimate the probability of an animal not being detected with a diagnostic test
}
