% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_plot_dist.R
\name{ra_plot_dist}
\alias{ra_plot_dist}
\title{Plot a distribution with descriptive statistics}
\usage{
ra_plot_dist(x, col = "red4", main = "")
}
\arguments{
\item{x}{Vector of the sampled values}

\item{col}{color used for the fill}

\item{main}{Main text used as a title for the plot}
}
\value{
Returns a ggplot histogram and ggtable arranged with ggarrange
}
\description{
A vector of numeric values is plotted as a histogram with descriptive statistics: min, max, standard deviation and mean
}
\examples{
# sample a distribution
x <- ra_sample(x = "Normal(100, 0.3)", 100)
ra_plot_dist(x)

}
