% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_plot_score.R
\name{ra_plot_score}
\alias{ra_plot_score}
\title{Plot the distribution using a gauge plot}
\usage{
ra_plot_score(x, main = "", brks = NULL)
}
\arguments{
\item{x}{vector of numbers}

\item{main}{Main title for the figure}

\item{brks}{breaks for the risk thresholds}
}
\value{
A plotly gauge figure
}
\description{
This function uses plotly to generate a gauge plot using a vector of numbers. Gauge plots provide a visual representation of a numeric magnitude
}
\examples{
# use one of the example models
m <- quantrra::OIRSA

ra_run(m = m$nodes, nsim = 100) \%>\% # run the model
  pull(P) \%>\% # extract the final probability
  ra_plot_score() # plot the score as a gauge plot
}
