% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_run_strat.R
\name{ra_run_strat}
\alias{ra_run_strat}
\title{Run a stratified model.}
\usage{
ra_run_strat(m, tbl, nsim, full = F, simplify = T)
}
\arguments{
\item{m}{Model file}

\item{tbl}{Table with parameters per strata}

\item{nsim}{Number of simulations}
}
\value{
data frame with mean, and 95% percentiles for each of the output nodes
}
\description{
The stratified model uses a different set of parameters for each starta with the intention to account for population heterogeneity. 
A strata represents different ways of grouping the target population. I.e. estimate the risk for different age groups, regions, or other groups to account for the population heterogeneity.
}
\examples{
# use one of the example models
m <- quantrra::asf_products
ra_run_strat(m = m$nodes, tbl = m$stratified, nsim = 10e3)

}
