% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_run_strat.R
\name{ra_run_strat}
\alias{ra_run_strat}
\title{Function to calculate the stratified risk}
\usage{
ra_run_strat(M, Tbl, nsim, full = F, simplify = T)
}
\arguments{
\item{M}{Model file}

\item{Tbl}{Table with parameters per strata}

\item{nsim}{Number of simulations}
}
\description{
Function to calculate the stratified risk
}
