% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_sample.R
\name{ra_sample}
\alias{ra_sample}
\title{Function to sample from different distributions}
\usage{
ra_sample(x, n, full = F)
}
\arguments{
\item{x}{Distribution specification, must specify the name of the distribution and parameters, e.g. Normal (10, 0.4) .Current distributions supported include: uniform, binomial, triangle, poisson, normal and pert.}

\item{n}{Number of observations to be sampled from the distribution}

\item{full}{TRUE/FALSE. If TRUE a list with the samples, and distribution specified will be returned, Only a vector of the observations will be returned otherwise.}
}
\description{
Function to sample from different distributions
}
\examples{
ra_sample(x = 'pert(0.01,  0.03, 0.64)', n = 10)
}
