% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_truncate.R
\name{ra_truncate}
\alias{ra_truncate}
\title{Function to truncate a distribution}
\usage{
ra_truncate(x, range, method = "ignore")
}
\arguments{
\item{x}{vector of values to truncate}

\item{range}{range of values to truncate at}

\item{method}{method for the truncation. default is 'ignore', which will ignore all the values out of the range. 'sub' will substitute the values out of the range for the min and max values. 'resample' will resample the values of out range from the values inside the range.}
}
\description{
Function to truncate a distribution
}
