% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_which.R
\name{ra_which}
\alias{ra_which}
\title{(experimental) Fit data to multiple distributions to see which one represents it best}
\usage{
ra_which(x, distributions)
}
\arguments{
\item{x}{Data to fit}

\item{dist}{Distribution specification}
}
\description{
This function fits a vector of numbers to multiple distributions and provides test statistics on the fit.
}
