% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRA.R
\name{RRA}
\alias{RRA}
\title{Function to run the stochastic probabilistic model using a QuantRRA table}
\usage{
RRA(M, nsim)
}
\arguments{
\item{M}{Model file}

\item{nsim}{Number of simulations}
}
\description{
Function to run the stochastic probabilistic model using a QuantRRA table
}
