% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRA_s.R
\name{RRA_s}
\alias{RRA_s}
\title{Function to calculate the stratified risk}
\usage{
RRA_s(M, Tbl, nsim, full = F, simplify = T)
}
\arguments{
\item{M}{Model file}

\item{Tbl}{Table with parameters per strata}

\item{nsim}{Number of simulations}
}
\description{
Function to calculate the stratified risk
}
