% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra_s.R
\name{rra_s}
\alias{rra_s}
\title{Function to calculate the stratified risk}
\usage{
rra_s(M, Tbl, nsim, full = F, simplify = T)
}
\arguments{
\item{M}{Model file}

\item{Tbl}{Table with parameters per strata}

\item{nsim}{Number of simulations}
}
\description{
Function to calculate the stratified risk
}
