% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhichDist.R
\name{WhichDist}
\alias{WhichDist}
\title{Function to sample from different distributions}
\usage{
WhichDist(x, distributions)
}
\arguments{
\item{x}{Data to fit}

\item{dist}{Distribution specification}
}
\description{
Function to sample from different distributions
}
